+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 3.3
+* Skipper changes with lot of improvement
 * Small engine changes as well as skipper changes
 * New function link_extract to parse more links
 * Small bug fixes